@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// Index
-class Index {
+class Index implements Model {
     /// Index Key.
     final String key;
     /// Index type.
@@ -11,14 +11,13 @@ class Index {
     /// Index attributes.
     final List attributes;
     /// Index orders.
-    final List orders;
+    final List? orders;
 
-    Index({
-        required this.key,
+    Index({        required this.key,
         required this.type,
         required this.status,
         required this.attributes,
-        required this.orders,
+this.orders,
     });
 
     factory Index.fromMap(Map<String, dynamic> map) {
