@@ -2,23 +2,19 @@ part of dart_appwrite;
 
     /// The Users service allows you to manage your project users.
 class Users extends Service {
-    Users(Client client): super(client);
+    Users(super.client);
 
     /// List Users
     ///
     /// Get a list of all the project's users. You can use the query params to
     /// filter your results.
     ///
-    Future<models.UserList> list({String? search, int? limit, int? offset, String? cursor, String? cursorDirection, String? orderType}) async {
+    Future<models.UserList> list({List<String>? queries, String? search}) async {
         final String path = '/users';
 
         final Map<String, dynamic> params = {
-            'search': search,
-'limit': limit,
-'offset': offset,
-'cursor': cursor,
-'cursorDirection': cursorDirection,
-'orderType': orderType,
+            'queries': queries,
+'search': search,
 
             
         };
@@ -34,13 +30,43 @@ class Users extends Service {
 
 
     }
-
     /// Create User
     ///
     /// Create a new user.
     ///
-    Future<models.User> create({required String userId, required String email, required String password, String? name}) async {
+    Future<models.User> create({required String userId, String? email, String? phone, String? password, String? name}) async {
         final String path = '/users';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'phone': phone,
+'password': password,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with Argon2 Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [Argon2](https://en.wikipedia.org/wiki/Argon2) algorithm. Use the [POST
+    /// /users](/docs/server/users#usersCreate) endpoint to create users with a
+    /// plain text password.
+    ///
+    Future<models.User> createArgon2User({required String userId, required String email, required String password, String? name}) async {
+        final String path = '/users/argon2';
 
         final Map<String, dynamic> params = {
             
@@ -62,7 +88,195 @@ class Users extends Service {
 
 
     }
+    /// Create User with Bcrypt Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [Bcrypt](https://en.wikipedia.org/wiki/Bcrypt) algorithm. Use the [POST
+    /// /users](/docs/server/users#usersCreate) endpoint to create users with a
+    /// plain text password.
+    ///
+    Future<models.User> createBcryptUser({required String userId, required String email, required String password, String? name}) async {
+        final String path = '/users/bcrypt';
 
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with MD5 Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [MD5](https://en.wikipedia.org/wiki/MD5) algorithm. Use the [POST
+    /// /users](/docs/server/users#usersCreate) endpoint to create users with a
+    /// plain text password.
+    ///
+    Future<models.User> createMD5User({required String userId, required String email, required String password, String? name}) async {
+        final String path = '/users/md5';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with PHPass Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [PHPass](https://www.openwall.com/phpass/) algorithm. Use the [POST
+    /// /users](/docs/server/users#usersCreate) endpoint to create users with a
+    /// plain text password.
+    ///
+    Future<models.User> createPHPassUser({required String userId, required String email, required String password, String? name}) async {
+        final String path = '/users/phpass';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with Scrypt Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [Scrypt](https://github.com/Tarsnap/scrypt) algorithm. Use the [POST
+    /// /users](/docs/server/users#usersCreate) endpoint to create users with a
+    /// plain text password.
+    ///
+    Future<models.User> createScryptUser({required String userId, required String email, required String password, required String passwordSalt, required int passwordCpu, required int passwordMemory, required int passwordParallel, required int passwordLength, String? name}) async {
+        final String path = '/users/scrypt';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'passwordSalt': passwordSalt,
+'passwordCpu': passwordCpu,
+'passwordMemory': passwordMemory,
+'passwordParallel': passwordParallel,
+'passwordLength': passwordLength,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with Scrypt Modified Password
+    ///
+    /// Create a new user. Password provided must be hashed with the [Scrypt
+    /// Modified](https://gist.github.com/Meldiron/eecf84a0225eccb5a378d45bb27462cc)
+    /// algorithm. Use the [POST /users](/docs/server/users#usersCreate) endpoint
+    /// to create users with a plain text password.
+    ///
+    Future<models.User> createScryptModifiedUser({required String userId, required String email, required String password, required String passwordSalt, required String passwordSaltSeparator, required String passwordSignerKey, String? name}) async {
+        final String path = '/users/scrypt-modified';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'passwordSalt': passwordSalt,
+'passwordSaltSeparator': passwordSaltSeparator,
+'passwordSignerKey': passwordSignerKey,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
+    /// Create User with SHA Password
+    ///
+    /// Create a new user. Password provided must be hashed with the
+    /// [SHA](https://en.wikipedia.org/wiki/Secure_Hash_Algorithm) algorithm. Use
+    /// the [POST /users](/docs/server/users#usersCreate) endpoint to create users
+    /// with a plain text password.
+    ///
+    Future<models.User> createSHAUser({required String userId, required String email, required String password, String? passwordVersion, String? name}) async {
+        final String path = '/users/sha';
+
+        final Map<String, dynamic> params = {
+            
+            'userId': userId,
+'email': email,
+'password': password,
+'passwordVersion': passwordVersion,
+'name': name,
+
+        };
+
+        final Map<String, String> headers = {
+            'content-type': 'application/json',
+
+        };
+
+        final res = await client.call(HttpMethod.post, path: path, params: params, headers: headers);
+
+        return models.User.fromMap(res.data);
+
+
+    }
     /// Get User
     ///
     /// Get a user by its unique ID.
@@ -86,7 +300,6 @@ class Users extends Service {
 
 
     }
-
     /// Delete User
     ///
     /// Delete a user by its unique ID, thereby releasing it's ID. Since ID is
@@ -114,7 +327,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Email
     ///
     /// Update the user email by its unique ID.
@@ -139,17 +351,15 @@ class Users extends Service {
 
 
     }
-
     /// Get User Logs
     ///
     /// Get the user activity logs list by its unique ID.
     ///
-    Future<models.LogList> getLogs({required String userId, int? limit, int? offset}) async {
+    Future<models.LogList> getLogs({required String userId, List<String>? queries}) async {
         final String path = '/users/{userId}/logs'.replaceAll('{userId}', userId);
 
         final Map<String, dynamic> params = {
-            'limit': limit,
-'offset': offset,
+            'queries': queries,
 
             
         };
@@ -165,7 +375,6 @@ class Users extends Service {
 
 
     }
-
     /// Get User Memberships
     ///
     /// Get the user membership list by its unique ID.
@@ -189,7 +398,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Name
     ///
     /// Update the user name by its unique ID.
@@ -214,7 +422,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Password
     ///
     /// Update the user password by its unique ID.
@@ -239,7 +446,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Phone
     ///
     /// Update the user phone by its unique ID.
@@ -264,7 +470,6 @@ class Users extends Service {
 
 
     }
-
     /// Get User Preferences
     ///
     /// Get the user preferences by its unique ID.
@@ -288,7 +493,6 @@ class Users extends Service {
 
 
     }
-
     /// Update User Preferences
     ///
     /// Update the user preferences by its unique ID. The object you pass is stored
@@ -315,7 +519,6 @@ class Users extends Service {
 
 
     }
-
     /// Get User Sessions
     ///
     /// Get the user sessions list by its unique ID.
@@ -339,7 +542,6 @@ class Users extends Service {
 
 
     }
-
     /// Delete User Sessions
     ///
     /// Delete all user's sessions by using the user's unique ID.
@@ -363,7 +565,6 @@ class Users extends Service {
 
 
     }
-
     /// Delete User Session
     ///
     /// Delete a user sessions by its unique ID.
@@ -387,7 +588,6 @@ class Users extends Service {
 
 
     }
-
     /// Update User Status
     ///
     /// Update the user status by its unique ID. Use this endpoint as an
@@ -413,7 +613,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Email Verification
     ///
     /// Update the user email verification status by its unique ID.
@@ -438,7 +637,6 @@ class Users extends Service {
 
 
     }
-
     /// Update Phone Verification
     ///
     /// Update the user phone verification status by its unique ID.
