@@ -1,7 +1,7 @@
 part of dart_appwrite.models;
 
 /// AttributeEmail
-class AttributeEmail {
+class AttributeEmail implements Model {
     /// Attribute Key.
     final String key;
     /// Attribute type.
@@ -17,8 +17,7 @@ class AttributeEmail {
     /// Default value for attribute when not provided. Cannot be set when attribute is required.
     final String? xdefault;
 
-    AttributeEmail({
-        required this.key,
+    AttributeEmail({        required this.key,
         required this.type,
         required this.status,
         required this.xrequired,
