@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Buckets List
-class BucketList {
+class BucketList implements Model {
     /// Total number of buckets documents that matched your query.
     final int total;
     /// List of buckets.
     final List<Bucket> buckets;
 
-    BucketList({
-        required this.total,
+    BucketList({        required this.total,
         required this.buckets,
     });
 
