@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Collections List
-class CollectionList {
+class CollectionList implements Model {
     /// Total number of collections documents that matched your query.
     final int total;
     /// List of collections.
     final List<Collection> collections;
 
-    CollectionList({
-        required this.total,
+    CollectionList({        required this.total,
         required this.collections,
     });
 
