@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Continent
-class Continent {
+class Continent implements Model {
     /// Continent name.
     final String name;
     /// Continent two letter code.
     final String code;
 
-    Continent({
-        required this.name,
+    Continent({        required this.name,
         required this.code,
     });
 
