@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Country
-class Country {
+class Country implements Model {
     /// Country name.
     final String name;
     /// Country two-character ISO 3166-1 alpha code.
     final String code;
 
-    Country({
-        required this.name,
+    Country({        required this.name,
         required this.code,
     });
 
