@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Countries List
-class CountryList {
+class CountryList implements Model {
     /// Total number of countries documents that matched your query.
     final int total;
     /// List of countries.
     final List<Country> countries;
 
-    CountryList({
-        required this.total,
+    CountryList({        required this.total,
         required this.countries,
     });
 
