@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Currencies List
-class CurrencyList {
+class CurrencyList implements Model {
     /// Total number of currencies documents that matched your query.
     final int total;
     /// List of currencies.
     final List<Currency> currencies;
 
-    CurrencyList({
-        required this.total,
+    CurrencyList({        required this.total,
         required this.currencies,
     });
 
