@@ -1,13 +1,13 @@
 part of dart_appwrite.models;
 
 /// Function
-class Func {
+class Func implements Model {
     /// Function ID.
     final String $id;
-    /// Function creation date in Unix timestamp.
-    final int $createdAt;
-    /// Function update date in Unix timestamp.
-    final int $updatedAt;
+    /// Function creation date in ISO 8601 format.
+    final String $createdAt;
+    /// Function update date in ISO 8601 format.
+    final String $updatedAt;
     /// Execution permissions.
     final List execute;
     /// Function name.
@@ -18,21 +18,20 @@ class Func {
     final String runtime;
     /// Function&#039;s active deployment ID.
     final String deployment;
-    /// Function environment variables.
-    final Map<String, dynamic> vars;
+    /// Function variables.
+    final List<Variable> vars;
     /// Function trigger events.
     final List events;
     /// Function execution schedult in CRON format.
     final String schedule;
-    /// Function next scheduled execution date in Unix timestamp.
-    final int scheduleNext;
-    /// Function next scheduled execution date in Unix timestamp.
-    final int schedulePrevious;
+    /// Function&#039;s next scheduled execution time in ISO 8601 format.
+    final String scheduleNext;
+    /// Function&#039;s previous scheduled execution time in ISO 8601 format.
+    final String schedulePrevious;
     /// Function execution timeout in seconds.
     final int timeout;
 
-    Func({
-        required this.$id,
+    Func({        required this.$id,
         required this.$createdAt,
         required this.$updatedAt,
         required this.execute,
@@ -51,18 +50,18 @@ class Func {
     factory Func.fromMap(Map<String, dynamic> map) {
         return Func(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
-            $updatedAt: map['\$updatedAt'],
+            $createdAt: map['\$createdAt'].toString(),
+            $updatedAt: map['\$updatedAt'].toString(),
             execute: map['execute'],
             name: map['name'].toString(),
             status: map['status'].toString(),
             runtime: map['runtime'].toString(),
             deployment: map['deployment'].toString(),
-            vars: map['vars'],
+            vars: List<Variable>.from(map['vars'].map((p) => Variable.fromMap(p))),
             events: map['events'],
             schedule: map['schedule'].toString(),
-            scheduleNext: map['scheduleNext'],
-            schedulePrevious: map['schedulePrevious'],
+            scheduleNext: map['scheduleNext'].toString(),
+            schedulePrevious: map['schedulePrevious'].toString(),
             timeout: map['timeout'],
         );
     }
@@ -77,7 +76,7 @@ class Func {
             "status": status,
             "runtime": runtime,
             "deployment": deployment,
-            "vars": vars,
+            "vars": vars.map((p) => p.toMap()),
             "events": events,
             "schedule": schedule,
             "scheduleNext": scheduleNext,
