@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Health Antivirus
-class HealthAntivirus {
+class HealthAntivirus implements Model {
     /// Antivirus version.
     final String version;
     /// Antivirus status. Possible values can are: `disabled`, `offline`, `online`
     final String status;
 
-    HealthAntivirus({
-        required this.version,
+    HealthAntivirus({        required this.version,
         required this.status,
     });
 
