@@ -1,12 +1,11 @@
 part of dart_appwrite.models;
 
 /// Health Queue
-class HealthQueue {
+class HealthQueue implements Model {
     /// Amount of actions in the queue.
     final int size;
 
-    HealthQueue({
-        required this.size,
+    HealthQueue({        required this.size,
     });
 
     factory HealthQueue.fromMap(Map<String, dynamic> map) {
