@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Health Status
-class HealthStatus {
+class HealthStatus implements Model {
     /// Duration in milliseconds how long the health check took.
     final int ping;
     /// Service status. Possible values can are: `pass`, `fail`
     final String status;
 
-    HealthStatus({
-        required this.ping,
+    HealthStatus({        required this.ping,
         required this.status,
     });
 
