@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Indexes List
-class IndexList {
+class IndexList implements Model {
     /// Total number of indexes documents that matched your query.
     final int total;
     /// List of indexes.
     final List<Index> indexes;
 
-    IndexList({
-        required this.total,
+    IndexList({        required this.total,
         required this.indexes,
     });
 
