@@ -1,13 +1,13 @@
 part of dart_appwrite.models;
 
 /// Membership
-class Membership {
+class Membership implements Model {
     /// Membership ID.
     final String $id;
-    /// Membership creation date in Unix timestamp.
-    final int $createdAt;
-    /// Membership update date in Unix timestamp.
-    final int $updatedAt;
+    /// Membership creation date in ISO 8601 format.
+    final String $createdAt;
+    /// Membership update date in ISO 8601 format.
+    final String $updatedAt;
     /// User ID.
     final String userId;
     /// User name.
@@ -18,17 +18,16 @@ class Membership {
     final String teamId;
     /// Team name.
     final String teamName;
-    /// Date, the user has been invited to join the team in Unix timestamp.
-    final int invited;
-    /// Date, the user has accepted the invitation to join the team in Unix timestamp.
-    final int joined;
+    /// Date, the user has been invited to join the team in ISO 8601 format.
+    final String invited;
+    /// Date, the user has accepted the invitation to join the team in ISO 8601 format.
+    final String joined;
     /// User confirmation status, true if the user has joined the team or false otherwise.
     final bool confirm;
     /// User list of roles
     final List roles;
 
-    Membership({
-        required this.$id,
+    Membership({        required this.$id,
         required this.$createdAt,
         required this.$updatedAt,
         required this.userId,
@@ -45,15 +44,15 @@ class Membership {
     factory Membership.fromMap(Map<String, dynamic> map) {
         return Membership(
             $id: map['\$id'].toString(),
-            $createdAt: map['\$createdAt'],
-            $updatedAt: map['\$updatedAt'],
+            $createdAt: map['\$createdAt'].toString(),
+            $updatedAt: map['\$updatedAt'].toString(),
             userId: map['userId'].toString(),
             userName: map['userName'].toString(),
             userEmail: map['userEmail'].toString(),
             teamId: map['teamId'].toString(),
             teamName: map['teamName'].toString(),
-            invited: map['invited'],
-            joined: map['joined'],
+            invited: map['invited'].toString(),
+            joined: map['joined'].toString(),
             confirm: map['confirm'],
             roles: map['roles'],
         );
