@@ -1,11 +1,10 @@
 part of dart_appwrite.models;
 
 /// Preferences
-class Preferences {
+class Preferences implements Model {
     final Map<String, dynamic> data;
 
-    Preferences({
-        required this.data,
+    Preferences({        required this.data,
     });
 
     factory Preferences.fromMap(Map<String, dynamic> map) {
