@@ -1,14 +1,13 @@
 part of dart_appwrite.models;
 
 /// Sessions List
-class SessionList {
+class SessionList implements Model {
     /// Total number of sessions documents that matched your query.
     final int total;
     /// List of sessions.
     final List<Session> sessions;
 
-    SessionList({
-        required this.total,
+    SessionList({        required this.total,
         required this.sessions,
     });
 
